Add KnownSkylarkUserDefinedRuleTypes
Summary: Add a cache for all user-defined rules that are encountered while parsing. We then will use this cache in the coercer to get params, implementation functions etc when we build SkylarkDescriptionArgs.  At the moment the lifetime is global, however in the future it may make more sense to make it per-cell and maybe reduce its lifetime further.

Reviewed By: sbalabanov

shipit-source-id: 1f90daa935