Add NativeLinkables
Summary:
I'm going to migrate the entire native linkables thing to stop passing
the CxxPlatform around. Part of that will be rewriting a ton of stuff
operating on NativeLinkableGroup to operate instead on a NativeLinkable
(still to be written) that has the CxxPlatform baked in.

I tried to do this in one blow, but there's way too much.

A large part of the complexity is in the NativeLinkableGroups utility
class. This change makes a simple dumb copy of that. Future diffs will
rewrite this class to operate on NativeLinkable instead of the group and
then I'll change things to call this one instead.

Reviewed By: bobyangyf

shipit-source-id: 23d1fac6f2