Don't use instanceof in consumers (haskell)
Summary:
Consumers of BuildRules should be interacting with them based on the
interfaces/providers that they expose and shouldn't be downcasting them
themselves and doing custom behavior.

This is particular important right now because I'm trying to change how
these things are implemented/exposed and these users are violating the
necessary encapsulation for doing that.

Reviewed By: philipjameson

shipit-source-id: 4bf87025a0