Introduce NativeLinkable
Summary:
NativeLinkable is basically the platform-locked version of
NativeLinkableGroup. This is what we want to migrate almost every user
of NativeLinkableGroup to. Currently, the users are only implicitly
locked to a cxxplatform by just being expected to propagate the cxx
platform to everything.

Updates NativeLinkables utility functions to be written in terms of
NativeLinkable instead of NativeLinkableGroup.

The goal/approach is explained a bit in the task

Reviewed By: jtorkkola

shipit-source-id: 4154d258f6