Introduce PlatformLockedNativeLinkableGroup
Summary:
This implements NativeLinkable in terms of just a pair of
NativeLinkableGroup and CxxPlatform and delegates to the
NativeLinkableGroup for all functions. This allows to add a simple
`getNativeLinkable(CxxPlatform cxxPlatform)` to all NativeLinkableGroup
implementations (with a crappy cache to cache the created locked
linkables).

Now that we have the `NativeLinkable` interface and a way to get one,
that allows us to start rewriting code to operate on just that
limited interface. Once that code is all rewritten, we'll move on to the
next step of implementing NativeLinkable with real implementations. It's
done in this order because several of the uses of the
NativeLinkableGroup interface are its own implementations. Since there's
too much to migrate in one shot, this weird state is needed to support
migration.

Reviewed By: jtorkkola

shipit-source-id: 394f773859