Inline part of NativeLinkableGroups.getTransitiveSharedLibraries
Summary:
Most of this function is operating on NativeLinkableGroups, but it
accepts just general unknown type of iterable for the first part. That
makes it really hard to convert to NativeLinkable (in fact basically
what it is doing is first a traversal on the unconfigured graph
followed by a traversal on the platform-specific graph).

This change inlines the non-NativeLinkableGroup part into the callers
(it's just a single function call).

Reviewed By: bobyangyf

shipit-source-id: 71b5540cea