Close channels before setting future
Summary:
CI picked up that the OutputMaterializer tests are flaky, part of the reason is that we set the future before closing the channels. The tests verify that channels are closed so we have a small race here.

Set all pending futures after the stream for large files or the batch read is complete (and channels are closed). Cleaned up a futures callback as well, and ended removing closeable since I don't think it actually helps with readability.

Reviewed By: mykola-semko

shipit-source-id: 49b9616f0b