Simplified build report endpoint's response
Summary:
Client's follow-up of diff: D15932414

Simplified the response's format. either the json `{ "uri" : <generated uri>}` or `{}` in case of success,
and an object with the field `error` in the case of a failure. If a failure happens, log all the response's body.

This helps in knowing what went wrong instead of failing silently and hiding error information from the client.

Reviewed By: bobyangyf

shipit-source-id: 402e8b19e3