Add KnownRuleTypes interface
Summary: Add a KnownRuleTypes interface and have KnownNativeRuleTypes implement it. This is used to generify the logic used in type coercion and DescriptionArg construction. Also add a method to KnownRuleTypesProvider that allows us to get the NativeRuleTypes in the few places that require knowledge specifically about native rules (where UserDefinedRules will not be applicable)

Reviewed By: bobyangyf

shipit-source-id: e00a5a1196