Use KnownUserDefinedRuleTypes
Summary:
Add KnownUserDefinedRuleTypes into KnownRuleTypes, and make sure we proxy to it properly. This happens when the identifier has a '//' in it (this buck.type value is only set from `rule()`'s callable), or if the `RuleType` says it was a user defined rule.

After this we'll be able to retrieve the user-defined-rules global state easily in the build file parser and elsewhere

Reviewed By: bobyangyf

shipit-source-id: aa980cb37d