Add basic implementation of SkylarkDescription
Summary: Create a simple context and call the user's implementation function when SkylarkDescription's ruleImpl method is called. Right now this just creates a (required) dummy action as we wait for a few other pieces to fall into place in the graph engine. These will also go away as more of the context object is filled in.

Reviewed By: bobyangyf

shipit-source-id: 80e9bfaffc