Add exclude-symlinks and case-sensitive options to globber
Summary:
In the future, I want to use WatchmanGlobber, matching case exactly and excluding symbolic links and directories. Add options to WatchmanGlobber to make this possible.

Aside from the EXCLUDE_DIRECTORIES option (which already existed as a boolean parameter), the options are currently unused. I plan to add callers in future diffs.

This diff should not change behavior.

Reviewed By: bobyangyf

shipit-source-id: 3ed74c8e6e