Fix stdout mangling when certain GC logging JVM args are used with Java 11
Summary:
`-XX:+PrintGCDetails` is deprecated in Java 11, and gets mapped to `-Xlog:gc*`. This arg logs GC info to stdout, which causes issues when we need to parse Buck's output programmatically. `-verbose:gc` is technically still supported, but also logs to stdout. Furthermore, `-Xloggc` is deprecated, and gets automatically replaced with `-Xlog:gc`, but the JVM stupidly writes the deprecation warning to stdout.

This change removes the first two args, and makes Buck do the replacement of `-Xloggc` to `-Xlog:gc` itself to ensure an unadulterated stdout.

Reviewed By: styurin

shipit-source-id: 456b3baaaa