Fix DeterministicManifestTest in Java11
Summary: The behavior of `java.util.jar.Manifest` has changed, such that it no longer considers the `\r\n`s it inserts as part of the 72 character output line limit. Instead of having different output formatting behavior in `DeterministicManifest` and two sets of tests based on the Java version, I'm just disabling the test that compares the output of the two in Java 11 for now. This is relatively simple code, unlikely to change.

Reviewed By: styurin

shipit-source-id: 1e5db3b147