Get report id from doctor's server response
Summary:
Extracting the rage report id from the rage url that the server's returns.

Also fixed & modified the corresponding test (`testJsonUpload()`) to guard against this breaking.
I'm pretty confident that test was wrong because it's not using `DoctorProtocolVersion.JSON` in the request, i.e. the test server returned a non-json response...

Reviewed By: philipjameson

shipit-source-id: 1bd2c86e44