Refactor BuildReportUploader
Summary:
`BuildReportUploader` contained logic for **uploading** but not  related to uploading a *build report*. I cut that logic to a separate base class `RequestUploader` that contains that logic.

It only exposes one method `uploadRequest` that can take an arbitrary `RequestBody` and optionally some extra queryParams to add to the url.

This reduces the possibility of making an error (e.g. forgetting to put the build id in the url), and also its need is going to be evident in the next diff, where I use it to create an uploader different from `BuildReportUploader`.

Reviewed By: philipjameson

shipit-source-id: f9320bef33