Upload key logger file to rage report endpoint
Summary: Uploading the key logger file necessary for the cache analysis, after it is written to disk. After the listener is closed, the file should be in disk so we try to upload it at that moment. All this runs as a background task so that this IO does not block the user.

Reviewed By: philipjameson

shipit-source-id: a6b16c5f6d