Introduce MultitenantQueryEnvironment.getRulesInBasePath()
Summary:
This is designed to make it easier to add a service endpoint for an IDE
that can request all of the rules defined in a build file. The service
can decide which attributes of `RawBuildRule` to expose.

If the service also had a schema so that it knew which rules where
"runnable," "testable," etc., then it could easily include the
"capabilities" associated with a rule so that the IDE could offer
appropriate actions associated with the rule.

For sure, this would be even more compelling if we had location
information available with each rule, but that is future work.

Reviewed By: sbalabanov

shipit-source-id: c155f8d7f2