Enabled buck to accept more arguments after kill command (buck kill -c a.b=1)
Summary:
Context: T46385547

The kill command only worked if kill was the only argument passed to buck (`argv[1:] == ["kill"]`) or the third, in case `--isolation_prefix` was provided (`len(argv) > 3 and argv[3:] == ["kill"] and argv[1] == "--isolation_prefix"`), and therefore, it would have been passed from the wrapper to the java buck daemon.

Created a list of wrapper arguments (only `--isolation_prefix` for now) and wrapper commands and their corresponding functions (just in case further changes to buck will need the wrapper to handle more commands in the future). Each argument from `argv` will be checked whether they are wrapper specific arguments and will be removed along with their parameters (only to be added after checking if, at the end of their removal, the element at index 1 in argv is a wrapper specific command).

Now, kill and killall should only be passed to the java daemon when the `--help` option is provided.

Reviewed By: KapJI

shipit-source-id: 9b1668fa4c