Extend android_binary to split configuration and filter targets by CPU filter constraints
Summary:
This change extends `android_binary` to support building an APK for a multiplaform (`android_platform`) with multiple nested platforms for native code. This allows to use configurable attributes (`select`) with targets that define native code.

`android_binary` is extended with a new attribute: `target_cpu_type_constraints`. This attribute defines a mapping between a CPU type (the same values as in `cpu_filters`) and a constraint value that defines the CPU type. This is a translation layer between Android binary's CPU values and constraints.

For example:

```
    target_cpu_type_constraints = {
        "armv7": "buck//config/constraints:arm",
        "x86": "buck//config/constraints:x86_64",
    },
```

To enable this logic the build needs to run with using `android_platform` as a target platform. Here's an example of a platform:

```
android_platform(
    name = "android-x86_64-arm",
    base_platform = ":android-os",
    native_platforms = [
        ":android-x86_64",
        ":android-arm",
    ],
)
```

In this case native targets from the transitive closure of the android binary will be created with configurations that use platforms specified in `native_platforms`. After that, during AndroidBinary build rule creation, the native targets are filtered and categorized using constraints specified in `target_cpu_type_constraints`.

For example, if "A" is a native target, it will be created with both ":android-x86_64" and ":android-arm" platforms in configurations. When grouping native targets by CPU type, `AndroidNativeLibsPackageableGraphEnhancer` would check both targets' platforms to match `buck//config/constraints:arm` and `buck//config/constraints:x86_64` and would group them accordingly.

Reviewed By: jtorkkola

shipit-source-id: d2bd688130