Upload RuleKeyDiagnostics Files if configured
Summary:
Add `BuildReportFileUploader` as a general way to upload files to the build report endpoint.

It differs from `ruleKeyLoggerFile` in that it's a different endpoint.
And it differs from `uploadDiffFile` in that the this one is uploading a from a path while the other one is uploading from an inputStream.

Uploading both `rule_key_diag_keys.txt` and `rule_key_diag_graph.txt` files after finishing writing if it's configured in the build report.

Reviewed By: philipjameson

shipit-source-id: 2c389dc6ca