Return implicit DefaultInfo based on artifacts used as outputs
Summary: If the user does not return a DefaultInfo from their impl function, and they do not have any "output" attributes, return a DefaultInfo with default_outputs specified as 'every artifact that was an output of an action'. In the future users will be able to just return a DefaultInfo object, or set those attributes to override this behavior.

Reviewed By: cjhopman

shipit-source-id: 40ef8896cf