Capture outputs in rule analysis graph
Summary: Per the previous suggestion, rather than maintaining the list of outputs used to construct a default DefaultInfo object inside of SkylarkRuleContextActions, move ActionRegistry to an interface, and implement a delegate pattern over the top of the real registry passed in via the context. This allows us to capture all outputs whenever an action is registered.

Reviewed By: bobyangyf

shipit-source-id: 27cf00adfe