Add SkylarkProviderInfo
Summary: Skylark's validation methods are a problem again. If we try to return `List<ProviderInfo>` from `SkylarkList.getContents()`, we get a compiler error about using a raw type rather than a specific type. We don't actually care about the generic type here, just that we've got a ProviderInfo. Add in a simple interface that all ProviderInfo objects can inherit from that we can use in validation logic in SkylarkDescription.

Reviewed By: cjhopman

shipit-source-id: 01795b4f49