Add simple name in toString() for provider keys
Summary: Right now this defaults to something that's not particularly useful/friendly for users (`BuiltInKey{infoClass=class com.facebook.buck.core.rules.providers.lib.DefaultInfo}`). For user defined rules we'd like to this to be easier to decipher, so just add a simple default `toString()` that shows the class name.

Reviewed By: cjhopman

shipit-source-id: a378e0ba63