Document ProviderInfoCollection.Builder.build() exceptions
Summary: Add the IllegalArgumentException to the documented contract when we call ProviderInfoCollection.Builder.build(). This makes it so that we only have to create intermediate objects when verifying user defined rule return values in the case where someone returns an invalid value. (We can catch this exception, then construct a more user friendly message using some intermediate objects in a place where we have full context)

Reviewed By: bobyangyf

shipit-source-id: f838834d35