Allow ProviderInfo to be returned from impl
Summary: Allow a list of ProviderInfo objects to be returned from users' implementation functions. This validates that all objects are of the right type, that there are no duplicate types, and uses the DefaultInfo provided if present in the return value, rather than constructing on from actions that were invoked. (This is a bit of a bear to test, as we don't have multiple outputs just yet)

Reviewed By: bobyangyf

shipit-source-id: c0dab7f201