Make attribute return Object
Summary: For some attribute types we want to return a non-coerced default value, as we don't have enough context to coerce it when we create the default value in the callable. A primary example is having a SourceListAttribute. The default value will be a list of strings that we then validate when we do the type coercion later.

Reviewed By: bobyangyf

shipit-source-id: 6ea7670ad9