Add SourceList attribute
Summary:
Add the SourceListAttribute. This takes a list of either files, or labels, and will make those dependencies for rules that use them. This is only a subset of what we could make this API, we're keeping it simple for now.

Note that this is a divergence from Bazel's "label_list'. Our goal in buck's Attributes is to use types, rather than a myriad of flags that require us to maintain a bunch of internal toggling logic.

Reviewed By: bobyangyf

shipit-source-id: 198eda6921