Add a transformWithDeps interface
Summary: Sometimes we need to do a bit of work in the analysis phase in order to get attributes' values into a useable state for users. Common examples are transforming a deps list into a list of ProviderCollectionInfo objects, or a list of sources into a list of artifacts. Add a simple interface here to do that transformation, we'll add it to the various attributes that require it later.

Reviewed By: bobyangyf

shipit-source-id: 69f7168adf