Run post-coercion transforms on specified attrs in SkylarkRuleContextAttr
Summary: Ensure that we do post-coercion transformations in the ctx.attrs object before returning values to users. This is used right now to turn attr.source_list attributes into a list of artifacts (rather than a list of SourcePaths). The loading cache is used so that we don't prematurely do these transformations if the user does not actually try to access the attribute (say if the attr is only used in one branch of an if statement).

Reviewed By: bobyangyf

shipit-source-id: 51bf18cfbc