Use Set instead of ImmutableSet where possible in com.facebook.buck.query.
Summary:
The widespread use of `ImmutableSet` in `com.facebook.buck.query` is
done based on the assumption that `ImmutableSet.copyOf()` is "free"
when the argument is also an `ImmutableSet`. Because we have Kotlin code
in `com.facebook.buck.multitenant` that interacts heavily with
`com.facebook.buck.query` that does not use Guava collections, this
assumption is no longer valid.

To that end, we replace `ImmutableSet` with `Set` where possible with
the understanding that a method that takes a `Set` should not expect to
be able to modify the `Set`: if it wants to modify it, then it should
create its own copy.

It turned out there were a handful of cases where we would get a
sequence of sets and create a union of them by creating an
`ImmutableSet.Builder` and invoking `addAll()` with each set in the
sequence. This creates an unnecessary copy when there is only one set in
the sequence, so this diff introduces
`com.facebook.buck.query.Unions.of()` to address this issue.

Reviewed By: sbalabanov

shipit-source-id: 165517f857