Eliminate the use of org.immutables in com.facebook.buck.query.
Summary:
The classes generated by org.immutables always make defensive copies.
There does not appear to be any way to disable this.
If Guava is on the classpath, then it uses `ImmutableSet.copyOf()` to
create the copy. Trying to disable the behavior using `jdkOnly=true`
ends up generating a method named `createUnmodifiableSet()` that
creates a copy via `new LinkedHashSet().addAll()`.

Again, because com.facebook.buck.multitenant prefers Kotlin Sets,
`ImmutableSet.copyOf()` is not free as it would be if the argument
passed to it were also an `ImmutableSet`.

Reviewed By: sbalabanov

shipit-source-id: 6376a3a701