Make Stampede's canonical cell name behavior not leak into internals of abstractions
Summary:
Stampede is using its own meaning for what a canonical cell name is and
then randomly changing cell's to use that canonical name. This leaks
into the apis that these underlying abstractions need to support and
exposes too much internal information.

Change the hack to work differently in a way that doesn't hook into
internals.

Reviewed By: styurin

shipit-source-id: 411c4bb6f2