Solved bug which hindered running buck without any arguments
Summary: When checking if `argv` contains commands which should be handled by the python wrapper, the `if` statement checked if the number of elements in `argv` is greater than 0, which is always true. Instead, it should've checked to see if it is greater than 1, so as to make sure there is at least one command to check against the python specific commands. Since running `buck` without any arguments passed that the if statement condition, `argv[1]` was still invalid and caused an exception.

Reviewed By: KapJI

shipit-source-id: 7300817d44