Add generic stringification method for skylark args
Summary:
Add a generic method to stringify objects that can be used as command line arguments (and that will be resolved at action execution time, rather than analysis time). This is like `CommandLineItem.expandToCommandLine()` in Bazel.

Note that we do this, rather than having type-specific `CommandLineArgs` objects (introduced in another diff) so that we are not allocating unnecessary one-time-use wrapper objects when one adds a long list of arguments in one go (e.g. with `ctx.actions.args().add_all([.....])`. As the only public interface to the entire args package will be via a few static factory methods, this seems to be a fine tradeoff for now. It may make sense to change it in the future if we add enough types that these helper methods become cumbersome.

Reviewed By: bobyangyf

shipit-source-id: b9b8c26b43