Add CommandLineArgs interface
Summary:
Add an interface to stick in front of collections of items that should be used as command line arguments such as strings, artifacts, labels, etc. We then use this interface at action execution time to resolve paths and the like

We want to have this basic interface so that we can have more complicated types that are more efficient behind the scenes in both the way the do transformations, and how they hold arguments. e.g. we can differentiate between a list of strings, and a list of other Arguments objects. In the latter case, we wouldn't need to copy the objects from the other Arguments objects, we could just hold the objects themselves.

Reviewed By: cjhopman

shipit-source-id: 25b227a7b2