Add ListCommandLineArgs interface
Summary: Add a simple CommandLineArgs implementation that handles lists of values. In the future, this will also take any transformation functions and the like that we may want to apply to all of the arguments. This allows us to be more efficient by not creating ScalarCommandLineArgs objects for each item.

Reviewed By: bobyangyf

shipit-source-id: 723050b3fe