Add AggregateCommandLineArgs interface
Summary: Add a CommandLineArgs interface that lets us iterate over multiple other CommandLineArgs objects. This is helpful in things like the run() action, where a user may provided a list of Args objects from providers, multiple invocations of args(), etc that may be very large. In those cases, we do not want to copy all of the items from the other CommandLineArgs objects, we just want to defer to them while iterating.

Reviewed By: bobyangyf

shipit-source-id: 308c990cc1