Introduce CanonicalCellName/CellNameResolver/NewCellPathResolver
Summary:
The way that we handle cell names and canonical cell names and cell
path lookups is so brittle (and wrong in places). We pass around
"canonical" cell names as Optional<String>, but we also accept
Optional<String> as a cell name for looking up paths and those two
aren't treated the same (ex. Optional.empty() maps to a different path
in every cell).

This introduce a few new things and creates implementations of them but
doesn't really use them anywhere yet.

It adds:

CanonicalCellName - Holds an Optional<String>. This is actually always
the canonical cell name.

NewCellPathResolver - This maps CanonicalCellName -> absolute Path.
This mapping is the same in every cell. Importantly, it isn't filtered
to the other cells mentioned in that cell's .buckconfig (which wouldn't
make sense because there's no reason to believe that the only cells
that you end up using are ones that you know about).

It's unfortunate that this is name "NewFoo", but I really will finish
migration to this, delete the old one and rename this to drop the "New".
Part of the reason to do it this way is that, as part of the migration,
I'll remove some of the uses of the old one (and so those will churn
only once in the approach from ""->delete as opposed to doing
""->"Legacy"->delete).

CellNameResolver - This maps cell-local names to CanonicalCellName. The
behavior of this is cell-specific (though, of course, the canonical
names are not). This is much more limited scope of cell-specific
behavior than the current CellPathResolver (and thus will, hopefully, be
much less of a source of complexity/bugs/confusion).

Reviewed By: sbalabanov

shipit-source-id: f30f49c4d6