PythonDslProjectBuildFileParser to restart buck.py process after an error
Summary:
Before resilient parser was introduced, this issue was hidden. When parsing error occurred, buck.py process crashed and terminated but because Buck terminated as well, the process was never reused. Resilient parser continues on parsing errors so it needs process to continue.
The easiest way seems to restart buck.py process (though it is very expensive). Alternative is to modify buck.py to not crash, but this might have unpredicted side effects.

Reviewed By: bobyangyf

shipit-source-id: 64937f6862