Ensure that hidden attributes are exposed and added as deps properly
Summary: We do not allow attributes starting with _ to be set by users (see https://docs.bazel.build/versions/master/skylark/rules.html#private-attributes-and-implicit-dependencies), however we need to make sure they're available in the implementation function, and get set properly as dependencies. Do this by exposing a `getPreCoercionDefaultValue()` on ParamInfo that keeps the old behavior on reflection based ParamInfo, but proxies Attribute.getPreCoercionDefaultValue() for skylark parameters. Default values are handled by Immutables for reflection based parameters, so it's fine to not set the values here for those types of constructor args.

Reviewed By: bobyangyf

shipit-source-id: 727096169f