Revert "Don't install APK-under-test if it is already present"
Summary:
D14684331 made it so that we don't install the APK-under-test if it is already present. That was mostly introduced to fix a problem we were seeing with running out of space on the device, when the real fix was D14803328.

In reality, this functionality doesn't really make a lot of sense - the APK-under-test should only already be present on the device if we used the `--attempt-uninstall-instrumentation-apk=false` option to leave it there, and in that case we can just not provide the apk-under-test path when we run another test.

The reason to remove it is (a) to simplify things, and (b) because it does take some time to do the check, so we should speed things up a little bit by removing it

Reviewed By: asp2insp

shipit-source-id: 569b3b1e6c