Fix error handling in makedirs()
Summary:
Fix the makedirs() function in buck_project.py to actually raise an
error if it fails to create the destination directory.  Previously this
incorrect condition would hide legitimate problems like EPERM or ENOENT.

The old code would only report errors if the directory already existed
but we got back an error other than EEXIST.  This generally would
result in buck failing later on since a directory it needed wasn't
present.  However, it made the problem harder to debug since we didn't
report the original reason why we failed to create the directory.

Reviewed By: mykola-semko

shipit-source-id: 53e39e4572