Fix prebuilt_cxx_library and haskell omnibus
Summary:
Recent refactoring incorrectly dropped the
forceLinkWholeForHaskellOmnibus implementation. This updates them to
pass along that information.

I chose to require it to be specified any time that setting
supportsOmnibusLinkingForHaskell is called. I think we could try it to
be only when you are setting that to "true", but this seems okay.

Reviewed By: mykola-semko

shipit-source-id: 2fc3ff0d96