Add attr.source()
Summary:
Add `attr.source()` for user defined rules. This creates an attribute that coerces the provided value to a single Artifact within the implementation function. An error is raised if a build target is provided that does not have exactly one output. If multiple outputs are allowed, use attr.source_list() with a singleton list.

Other things in https://docs.bazel.build/versions/master/skylark/lib/attr.html#label may or may not be added, but the leading candidates would be `executable` and `providers` (e.g. to make sure that an output came from a specific type of rule)

Reviewed By: bobyangyf

shipit-source-id: 47666f5eb2