Add list converter that supports generics
Summary:
We're going to want to have a function to validate that we got a list of `Provider<?>`s in a future diff. The built in SkylarkList validation methods don't let us do this because `Provider` is a generic class, and we can't get `Class<Provider<?>>` to pass in. So add a simple method (with examples of what we're trying to handle) to let us do:
```
SkylarkList<?> skyList = // Provided by user
ImmutableList<Provider<?>> providers = BuckSkylarkTypes.toJavaList(skyList, Provider.class, null));
```

Reviewed By: bobyangyf

shipit-source-id: c87b8edf4b