Add provider restrictions to several attrs
Summary:
Add the ability to ensure that specified dependencies have a given set of Providers available. This can be used to make sure that one is only depending on certain types of rules (e.g. make sure to only depend on things that export C++ provider infos)

This has been added to: source, source_list, dep, dep_list. source and source_list still are transformed into artifacts, however, this can be used as a crude form of guaranteeing that only sources of a certain "type" are provided.

Reviewed By: bobyangyf

shipit-source-id: 14fceb8d2a