Delete misleading overload of DepsAwareTask.of
Summary:
DepsSupplier.of's signature implies that the function should return deps, not prereqs:

  public static <U> DepsSupplier<U> of(ThrowingSupplier<ImmutableSet<U>, Exception> depSupplier) {

However, its implementation sets this.prereqSupplier. This function's interface is misleading. I think calling code is easier to understand if it's more explicit whether dependencies are deps or rereqs.

Inline this overload of DepsSupplier.of into its callers and delete it. Also, add an overload of DepsSupplier.of with no arguments.

This diff should not change behavior.

Reviewed By: bobyangyf

shipit-source-id: 71fe7013ee