Parse BuckFixSpec from BuckFixFile
Summary:
Parsing a `BuckFixSpec` is just deserializing the json file stored in logs.
Added a way to run that parser directly from the `FixCommandHandler`, But kept the way we write the buckFixSpec to a temporary file indicated by the python wrapper. We might want to eventually change that (i.e. remove passing back and forth `--fix-spec-file` from the wrapper to the client, and only invoke the fix script with the path to the file located inside the log directory).

Reviewed By: philipjameson

shipit-source-id: 104ce69431