Add kotlin prefix to android library type
Summary: Right now, for an Android library written in Kotlin you just add `language = "KOTLIN"` to your `android_library`. This means that we don't get a different rule type, so we can't tell which targets are being built using Kotlin and which are being built using Java. This diff makes it so that a target built using Kotlin has a rule type of `kotlin_android_library`.

Reviewed By: asp2insp

shipit-source-id: 54f8594043