Also check whether exported deps of a dependency are used for java.unused_dependencies_action
Summary:
Bit torn over this one, as technically we should just be importing the dep that actually used directly, and not using the `exported_deps`. However, without this change, we give the user incorrect information that they should remove the dep, which then breaks their build.

There are a couple of options in the case where only the exported deps are actually used:
1. We could tell them to remove this dep, and tell them which dep(s) they need to add. However, I think there's still a possibility that this might cause problems with visibility (the exported dep C will be visible to the target B that is exporting it, but not necessarily to target A which depends on B, I think?)
2. We can go with this and allow targets that are not used but have exported deps that are used.

If my intuition on the visibility is correct, I think 2 is probably the way to way, but if not then we should probably do 1 instead. Thoughts?

Reviewed By: asp2insp

shipit-source-id: 898a15dd2d