Remove combinedProject path in projectv2
Summary: For projectv2, we only want to support the project directory structure and single target that the `combined-project` flag handles. However, this code path only sort of works, and it has a few issues. As a result, this deletes the flag from this code path and defaults to the original code path. I will use the deleted code as a reference going forward; however, right now `buck project --combined-project` generates a broken Xcode project with build errors, and thus we should use the one code path that we know works as the basis going forward.

Reviewed By: williamtwilson

shipit-source-id: 12e1604b23