Add link group parameters to LinkableCxxConstructorArg
Summary:
Adds rule parameters for link groups, namely link_group and link_group_map. In addition, add the necessary type coercers to be able to parse the new fields in BUCK files.

NOTE: The parameters had to be added to `LinkableCxxConstructorArg` as that's how we can pass information from the Apple descriptions to underlying Cxx rule factories.

Reviewed By: williamtwilson

shipit-source-id: da1338875d