Add support for link groups
Summary:
In this diff, we add support for link groups. This is done by introducing `LinkableListFilterFactory` which can create a filter using `CxxBuckConfig`, arguments and a target graph.

As we use the config option defined in the previous diff, this diff introduced no perf regression as no additional code would run as part of linking.

NOTE: We require access to the target graph, by definition, to be able to compute which link group a particular target belongs to.

Reviewed By: williamtwilson

shipit-source-id: 83dce84bfb