Implement caching for URLClassPath
Summary:
The previous iteration of this code used the `UrlClassLoader#addUrl` function to append URLs to the existing classloader. This disables the internal JVM cache which maps class names to their location in jars (in fact, any modification of the classpath will disable the cache). Without this cache, each class lookup takes linear time in the total number of jars on the classpath.

For JUnit tests, Buck puts the full transitive dependency tree onto the classpath, meaning that for a large codebase with lots of small modules, the number of jars can be large. This can cause significant slowdowns for tests.

`CachingClassPath` implements a very simplistic cache where each package name is mapped to up to two smaller URLClassPath instances. Both of these smaller instances are queried before falling back to a linear search. In testing this reduced runtime for large classpath tests by up to 30%.

The 2-slot cache was chosen because of the common pattern for unit tests: a `java_test` and the `java_library` under test will create two separate jars which share the same package name.

Reviewed By: IanChilds

shipit-source-id: 6cf8d3ce57