Fix Xcode 11 Swift + macOS Compatibility
Summary:
Xcode 11 changed the way the swift libraries were organized inside the toolchain so Buck needs to be updated to accomodate. Now we dynamically look inside the toolchain for the first folder that starts with swift and contains libSwiftCore.dylib (per the platform targeted).

It's quite possible that this folder structure will change in the future, so this could always be revisited. In addition... I don't know how specific versions will be handled (i.e. in Xcode 11 there's swift-5.0, but perhaps that may change to swift-5.1 or both swift-5.0 and swift-5.1 and we'll need to change the way specific versioning is handled.

This diff also applies the suggested fix of disabling static linking on macOS as it's no longer supported / needed.

Reviewed By: milend

shipit-source-id: 3b2b6666b4